Phase 7: Rewrite ProductionService with HTTP clients + fix all build errors

+++ BackendAPI/Dtos/Production/ProductionPlanningInfoDto.cs
@@ -7,7 +7,7 @@ namespace BackendAPI.Dtos.Production
         public decimal RemainingQuantity { get; set; } 
         
         // Limits
-        public int MachineDailyCapacity { get; set; }
+        public decimal MachineDailyCapacity { get; set; }
         public decimal MaxPossibleByMaterial { get; set; }
         public string LimitingMaterial { get; set; } = string.Empty;
     }
